Comment out "Don't show mobile data options if not supported"
Apparently the check on telephony to decide if it has mobile
data "MAY" be incorrect. Tilapia has mobile data but no telephony
and as such it was checked as no mobile data for tilapia, which
is incorrect.

This is a dirty workaround of commenting out the check.
Will find a proper solution.