SystemUI: option to toggle dimming of the notification shade (2/2)
Allow the user to enable/disable dimming of the notification shade
pulldown. Check isHighEndGfx() to respect the default setting that
the device currently supports.

Users on low-end device may want to have this feature, while users
on high-end devices may want to disable it (possible battery life
savings and what not).

Change-Id: I661254ec7ae1548707d94f5d41efe877554c6023